+++ include/riscv_pkg.sv
@@ -354,26 +354,55 @@ package riscv;
         CSR_MHARTID        = 12'hF14,
         CSR_MCYCLE         = 12'hB00,
         CSR_MINSTRET       = 12'hB02,
+        // Performance counters (Machine Mode)
+        CSR_ML1_ICACHE_MISS = 12'hB03,  // L1 Instr Cache Miss
+        CSR_ML1_DCACHE_MISS = 12'hB04,  // L1 Data Cache Miss
+        CSR_MITLB_MISS      = 12'hB05,  // ITLB Miss
+        CSR_MDTLB_MISS      = 12'hB06,  // DTLB Miss
+        CSR_MLOAD           = 12'hB07,  // Loads
+        CSR_MSTORE          = 12'hB08,  // Stores
+        CSR_MEXCEPTION      = 12'hB09,  // Taken exceptions
+        CSR_MEXCEPTION_RET  = 12'hB0A,  // Exception return
+        CSR_MBRANCH_JUMP    = 12'hB0B,  // Software change of PC
+        CSR_MCALL           = 12'hB0C,  // Procedure call
+        CSR_MRET            = 12'hB0D,  // Procedure Return
+        CSR_MMIS_PREDICT    = 12'hB0E,  // Branch mis-predicted
+        CSR_MSB_FULL        = 12'hB0F,  // Scoreboard full
+        CSR_MIF_EMPTY       = 12'hB10,  // instruction fetch queue empty
+        CSR_MHPM_COUNTER_17 = 12'hB11,  // reserved
+        CSR_MHPM_COUNTER_18 = 12'hB12,  // reserved
+        CSR_MHPM_COUNTER_19 = 12'hB13,  // reserved
+        CSR_MHPM_COUNTER_20 = 12'hB14,  // reserved
+        CSR_MHPM_COUNTER_21 = 12'hB15,  // reserved
+        CSR_MHPM_COUNTER_22 = 12'hB16,  // reserved
+        CSR_MHPM_COUNTER_23 = 12'hB17,  // reserved
+        CSR_MHPM_COUNTER_24 = 12'hB18,  // reserved
+        CSR_MHPM_COUNTER_25 = 12'hB19,  // reserved
+        CSR_MHPM_COUNTER_26 = 12'hB1A,  // reserved
+        CSR_MHPM_COUNTER_27 = 12'hB1B,  // reserved
+        CSR_MHPM_COUNTER_28 = 12'hB1C,  // reserved
+        CSR_MHPM_COUNTER_29 = 12'hB1D,  // reserved
+        CSR_MHPM_COUNTER_30 = 12'hB1E,  // reserved
+        CSR_MHPM_COUNTER_31 = 12'hB1F,  // reserved
+        // Cache Control (platform specifc)
         CSR_DCACHE         = 12'h701,
         CSR_ICACHE         = 12'h700,
-
+        // Triggers
         CSR_TSELECT        = 12'h7A0,
         CSR_TDATA1         = 12'h7A1,
         CSR_TDATA2         = 12'h7A2,
         CSR_TDATA3         = 12'h7A3,
         CSR_TINFO          = 12'h7A4,
-
         // Debug CSR
         CSR_DCSR           = 12'h7b0,
         CSR_DPC            = 12'h7b1,
         CSR_DSCRATCH0      = 12'h7b2, // optional
         CSR_DSCRATCH1      = 12'h7b3, // optional
-
-        // Counters and Timers
+        // Counters and Timers (User Mode - R/O Shadows)
         CSR_CYCLE          = 12'hC00,
         CSR_TIME           = 12'hC01,
         CSR_INSTRET        = 12'hC02,
-        // Performance counters
+        // Performance counters (User Mode - R/O Shadows)
         CSR_L1_ICACHE_MISS = 12'hC03,  // L1 Instr Cache Miss
         CSR_L1_DCACHE_MISS = 12'hC04,  // L1 Data Cache Miss
         CSR_ITLB_MISS      = 12'hC05,  // ITLB Miss
@@ -387,7 +416,22 @@ package riscv;
         CSR_RET            = 12'hC0D,  // Procedure Return
         CSR_MIS_PREDICT    = 12'hC0E,  // Branch mis-predicted
         CSR_SB_FULL        = 12'hC0F,  // Scoreboard full
-        CSR_IF_EMPTY       = 12'hC10   // instruction fetch queue empty
+        CSR_IF_EMPTY       = 12'hC10,  // instruction fetch queue empty
+        CSR_HPM_COUNTER_17 = 12'hC11,  // reserved
+        CSR_HPM_COUNTER_18 = 12'hC12,  // reserved
+        CSR_HPM_COUNTER_19 = 12'hC13,  // reserved
+        CSR_HPM_COUNTER_20 = 12'hC14,  // reserved
+        CSR_HPM_COUNTER_21 = 12'hC15,  // reserved
+        CSR_HPM_COUNTER_22 = 12'hC16,  // reserved
+        CSR_HPM_COUNTER_23 = 12'hC17,  // reserved
+        CSR_HPM_COUNTER_24 = 12'hC18,  // reserved
+        CSR_HPM_COUNTER_25 = 12'hC19,  // reserved
+        CSR_HPM_COUNTER_26 = 12'hC1A,  // reserved
+        CSR_HPM_COUNTER_27 = 12'hC1B,  // reserved
+        CSR_HPM_COUNTER_28 = 12'hC1C,  // reserved
+        CSR_HPM_COUNTER_29 = 12'hC1D,  // reserved
+        CSR_HPM_COUNTER_30 = 12'hC1E,  // reserved
+        CSR_HPM_COUNTER_31 = 12'hC1F  // reserved
     } csr_reg_t;
 
     localparam logic [63:0] SSTATUS_UIE  = 64'h00000001;